Organized Code
Will be helpfull for add storybook for testing - accessibility
