+++ src/index.ts
@@ -1,5 +1,6 @@
-import { BiteConsent } from './BiteConsent'
-import Position, { CustomPosition } from './Position'
-import useBiteConsent from './useBiteConsent'
+import { BiteConsent } from './components/BiteConsent'
+import Position, { CustomPosition } from './utils/Position'
+import useBiteConsent from './hooks/useBiteConsent'
 
 export { BiteConsent, CustomPosition, Position, useBiteConsent }
+
